backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
scipy==1.13.1
rasterio==1.3.10
requests==2.32.3
pillow==10.4.0
matplotlib==3.9.4
anthropic==0.39.0
pycma==3.4.0
astral==3.2
pydantic==2.9.0
python-multipart==0.0.9
